requirements.txt
python-dotenv
requests
requests_oauthlib